Extracted 3rd party libs. Added multiple file systems and tests
