+++ Tests/Updater/FileSystemTest.php
@@ -2,33 +2,32 @@
 
 namespace MyBuilder\Cronos\Tests\Formatter;
 
-use MyBuilder\Cronos\Updater\FileSystem;
+use MyBuilder\Cronos\Updater\SymfonyFileSystem;
+use MyBuilder\Cronos\Updater\StandardFileSystem;
 
-/**
- * FileSystemTest
- */
 class FileSystemTest extends \PHPUnit_Framework_TestCase
 {
     /**
-     * @var FileSystem
+     * @test
+     * @dataProvider availableFileSystems
      */
-    private $object;
-
-    public function setUp()
+    public function shouldCreateTempFile($fileSystem)
     {
-        $this->object = new FileSystem();
-    }
-
-    public function testFilesystem()
-    {
-        $filename = $this->object->createTempFile('test', 'test content');
+        $filename = $fileSystem->createTempFile('test', 'test content');
 
         $this->assertTrue(file_exists($filename));
         $this->assertEquals('test content', file_get_contents($filename));
 
-        $this->object->removeFile($filename);
+        $fileSystem->removeFile($filename);
 
         $this->assertFalse(file_exists($filename));
     }
+
+    public function availableFileSystems()
+    {
+        return array(
+            array(new SymfonyFileSystem),
+            array(new StandardFileSystem)
+        );
+    }
 }
-
