+++ Updater/ProcessRunner.php
@@ -2,12 +2,10 @@
 
 namespace MyBuilder\Cronos\Updater;
 
-use Symfony\Component\Process\Process;
-
 /**
  * Responsible for running processes
  */
-class ProcessRunner
+interface ProcessRunner
 {
     /**
      * Run the given command
@@ -16,11 +14,5 @@ class ProcessRunner
      *
      * @return boolean
      */
-    public function run($command)
-    {
-        $process = new Process($command);
-        $process->run();
-
-        return $process->isSuccessful();
-    }
+    public function run($command);
 }
