+++ composer.json
@@ -18,8 +18,10 @@
     ],
     "require": {
         "php": ">=5.3.3",
-        "symfony/process": ">=2.2,<=2.4",
-        "symfony/filesystem": ">=2.2,<=2.4"
+        "symfony/process": "2.2.*"
+    },
+    "suggest": {
+        "symfony/filesystem": "Allows using Symfony Filesystem"
     },
     "require-dev": {
         "phpunit/phpunit": "3.7.*",
